@@TRANCOUNT
This is a system function in SQL Server that returns the number of active transactions for the current session.

If @@TRANCOUNT = 0, there are no active transactions.

If @@TRANCOUNT > 0, it indicates how many nested transactions are currently open.

AS ActiveTransactions

Gives a friendly name to the output column for easier reading.